+++ test/Peast/Syntax/ES2022/files/ClassStaticBlock/ClassStaticBlock3.js
@@ -1,0 +1,8 @@
+class foo{
+	static {
+		this.test = 1
+	}
+	static {
+		this.test2 = 2
+	}
+}